chore: move block body to primitives
