+++ crates/net/eth-wire/src/types/blocks.rs
@@ -1,9 +1,7 @@
 //! Implements the `GetBlockHeaders`, `GetBlockBodies`, `BlockHeaders`, and `BlockBodies` message
 //! types.
 use reth_codecs::derive_arbitrary;
-use reth_primitives::{
-    Block, BlockHashOrNumber, Header, HeadersDirection, TransactionSigned, Withdrawal, H256,
-};
+use reth_primitives::{BlockBody, BlockHashOrNumber, Header, HeadersDirection, H256};
 use reth_rlp::{RlpDecodable, RlpDecodableWrapper, RlpEncodable, RlpEncodableWrapper};
 
 #[cfg(feature = "serde")]
@@ -65,35 +63,6 @@ pub struct GetBlockBodies(
 impl From<Vec<H256>> for GetBlockBodies {
     fn from(hashes: Vec<H256>) -> Self {
         GetBlockBodies(hashes)
-    }
-}
-
-// TODO(onbjerg): We should have this type in primitives
-/// A response to [`GetBlockBodies`], containing bodies if any bodies were found.
-///
-/// Withdrawals can be optionally included at the end of the RLP encoded message.
-#[derive_arbitrary(rlp, 10)]
-#[derive(Clone, Debug, PartialEq, Eq, Default, RlpEncodable, RlpDecodable)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[rlp(trailing)]
-pub struct BlockBody {
-    /// Transactions in the block
-    pub transactions: Vec<TransactionSigned>,
-    /// Uncle headers for the given block
-    pub ommers: Vec<Header>,
-    /// Withdrawals in the block.
-    pub withdrawals: Option<Vec<Withdrawal>>,
-}
-
-impl BlockBody {
-    /// Create a [`Block`](reth_primitives::Block) from the body and its header.
-    pub fn create_block(&self, header: Header) -> Block {
-        Block {
-            header,
-            body: self.transactions.clone(),
-            ommers: self.ommers.clone(),
-            withdrawals: self.withdrawals.clone(),
-        }
     }
 }
