+++ crates/net/network/src/fetch/mod.rs
@@ -2,14 +2,14 @@
 
 use crate::{message::BlockRequest, peers::PeersHandle};
 use futures::StreamExt;
-use reth_eth_wire::{BlockBody, GetBlockBodies, GetBlockHeaders};
+use reth_eth_wire::{GetBlockBodies, GetBlockHeaders};
 use reth_interfaces::p2p::{
     error::{EthResponseValidator, PeerRequestResult, RequestError, RequestResult},
     headers::client::HeadersRequest,
     priority::Priority,
 };
 use reth_network_api::ReputationChangeKind;
-use reth_primitives::{Header, PeerId, H256};
+use reth_primitives::{BlockBody, Header, PeerId, H256};
 use std::{
     collections::{HashMap, VecDeque},
     sync::{
